feat: 清除nt痕迹， 更改的apis目录结构
