+++ common/vue/src/apis/auth.api.js
@@ -1,10 +1,10 @@
 /*
  * @Author: benaozhi
  * @Date: 2020-01-03 18:29:42
- * @LastEditTime: 2020-08-03 19:00:59
+ * @LastEditTime: 2020-09-04 14:51:32
  * @Description:
  */
-import Axios, { AxiosConfig } from '@/apis/services/axios'
+import Axios, { AxiosConfig } from '@/apis/http/strict-axios'
 
 // 登录
 export function login({ username, password }) {
