+++ packages/vue-elementui-docker-boilerplate/src/components/system/form/index.js
@@ -1,11 +1,11 @@
 /*
  * @Author: Cphayim
  * @Date: 2019-05-19 22:27:16
- * @LastEditTime: 2020-07-07 09:02:09
+ * @LastEditTime: 2020-09-03 02:11:57
  * @Description: 组件、结构化数据类聚合出口
  */
 
-// 导出 NTForm 组件
+// 导出 VRNForm 组件
 export { default as VRNForm } from './vrn-form.vue'
 
 // 导出所有表单控件组件
